+++ Change_Log.md
@@ -2,11 +2,15 @@
 
 ## Minor Release - Navbar & Theme Wrapping
 ### v0.1.0.0
-Theme Service - Add the framework for an applications / components / elements to easily toggle themes.
+<br />Theme Service - Add the framework for an applications / components / elements to easily toggle themes.
 Added some initial framework to create a baseline to test the theme against.  This
 would include a navbar, sidebar, and dashboard framework.  These components will be more
 fully devloped in future development
-Navbar - A component for the display of branding, menus and general application control.  The configuration of this component is done so with the complementary Navbar Service.  The navbar service allows for the customization of the Brand image, Brand text & menu contents
+<br />Navbar - A component for the display of branding, menus and general application control.  The configuration of this component is done so with the complementary Navbar Service.  The navbar service allows for the customization of the Brand image, Brand text & menu contents.  Navbar has ability to
+drop additional component as a heads-up display tailorable to user preference.
+<br />ToolFrame - a base class for class projection.  There are intended to be multiple modes at which
+a component can be displayed based upon where it is in the UI.  Currently, this frame support adding
+of mini-tools.
 <br />Added Content:
 - Theme Service
 - Navbar Service
@@ -14,6 +18,7 @@ Navbar - A component for the display of branding, menus and general application 
 - Dashboard
 - Sidebar
 - Login Page
+- Tool Frame
 
 ## Empty Release
 ### v0.0.0.0